Fix #1861 - Merge desktop and mobile headers, fixing bottom Navbar issue on iPhone
